chore(ux): changes StringVal and BoolVal into String and Bool to avoid confusion. (#48)
Functions config.StringVal and config.BoolVal allow user to easily return values to be used in the agent config, however and more so because this is a config, they could give the sense that we are retrieving the value for a config with a given key. Hence this change removes the suffix 'Val' to avoid such confusion. Thanks to @marc-gr for the hint.